Added in further GUI stuff such as the map + partial functionality.
The map plotting functionality -> The locations
of each state is taken from a text file, and given that a given state 
appears in the derived records, the alien is drawn. The size of the
alien is determined by the deviation of the UFO sightings frequency from
the average of all records being viewed.