Changes made to locations.txt file to include all states.
All mainland states of North America have been added to the locations 
file. Also, i have added the names of the states as well as the 
abbreviations. The changes required to facilitate this change in the 
locations file has been added to the class Map.Filereader to allow it to
interpret the data as expected.