Changes to the general layout of the project
I have made changes to the layout of the project interms of separating
functionality from the Process class. The process class has been renamed
to API - it now only handles trivial API related stuff such as getting
acknowledgements, last updates etc. It does implicity get the data to be
used i.e. it calls another class to do this, then returns the data. This
is because the data retreival process involves caching which would make
the class over conjested if i were to leave everything in one class. So
i divided the data retreieval into a separate class.